Most of this is sleep deprived, tear fueled code.¯\_(ツ)_/¯ But The programme will now ask for the users consent before deleting records from the database.＼（＾０＾）ノ The exception being frmError and frmSolution as the delete buttons had the same amount of code as I have self worth. ( ° ͜ʖ °) The user will now be asked if they are sure they want to abandon the hub forms and sign out, leaving them alone.ಥ﹏ಥ Also added pretty icons to the pop-up messageboxes (╯°□°）╯︵ ┻━┻

+++ Final Project/The Error Application/frmYear.cs
@@ -62,24 +62,38 @@ namespace The_Error_Application
 
         private void btnDelete_Click(object sender, EventArgs e)
         {
-            Year year = new Year();
-
-            year.YearID = int.Parse(dgvYear.SelectedRows[0].Cells["YearID"].Value.ToString());
-
-            int x = bll.DeleteYear(year);
-            if (x > 0)
+            //Setting up a Yes No message box for user confirmation
+            MessageBoxButtons buttons = MessageBoxButtons.YesNo;
+            DialogResult result = MessageBox.Show("Are you sure you wish to delete?", "Confirmation", buttons, MessageBoxIcon.Question);
+            if (result == DialogResult.Yes)
             {
-                MessageBox.Show(x + " Deleted!!");
-            }
-            else
-            {
-                MessageBox.Show(x + " Deleted!!");
+
+                Year year = new Year();
+
+                year.YearID = int.Parse(dgvYear.SelectedRows[0].Cells["YearID"].Value.ToString());
+
+                int x = bll.DeleteYear(year);
+                if (x > 0)
+                {
+                    MessageBox.Show(x + " Deleted!!");
+                }
+                else
+                {
+                    MessageBox.Show(x + " Deleted!!");
+                }
             }
         }
 
         private void btnDisplay_Click(object sender, EventArgs e)
         {
             dgvYear.DataSource = bll.GetYear();
+        }
+
+        private void exitToolStripMenuItem_Click(object sender, EventArgs e)
+        {
+            frmAdminHub frmAdminHub = new frmAdminHub();
+            frmAdminHub.Show();
+            this.Hide();
         }
     }
 }
